Updates for deep test case
Changes to support the corner case where an indy player caps a
territory, is invited to a group, joins the group (bringing the capped
territory to the joined group ownership), then leaves the group (leaving
the capped territory behind), and then leaves/rejoins the server.  When
he rejoins, the territory he capped re-attaches to him, and the group he
left previously loses the territory cap.  These changes support that
functionality, including updating the maps of players in the group the
re-joining player previously left.